Add difflib to autocorrect parsed names

+++ cogs/Raiding/vc_select.py
@@ -10,6 +10,7 @@ class VCSelect:
         self.client = client
         self.ctx = ctx
         self.headcount = headcount
+        self.parse = parse
         title = "Headcount Setup" if headcount else "Lock Selection" if lock else "Unlock Selection" if unlock else "Cleaning Selection" if\
             clean else "Parsing Setup" if parse else "AFK-Check Setup"
         description = "Please choose what channel you'd like to start this headcount in." if headcount else\
@@ -26,7 +27,8 @@ class VCSelect:
 
 
     async def start(self):
-        await self.ctx.message.delete()
+        if not self.parse:
+            await self.ctx.message.delete()
 
         if self.ctx.channel == self.guild_db.get(sql.gld_cols.raidcommandschannel):
             s = ""
